Rename DisplayChainedError to DisplayError and add it to MullvadTypes
